requirements.txt
torch
torchvision
timm
fastapi
uvicorn
pillow
numpy
matplotlib
huggingface_hub
